+++ docs/job_controller/overview.rst
@@ -41,10 +41,10 @@ In calling function::
     USE_BALSAM = False
 
     if USE_BALSAM:
-        from libensemble.balsam_controller import BalsamJobController
+        from libensemble.executors.balsam_controller import BalsamJobController
         jobctrl = BalsamJobController()
     else:
-        from libensemble.mpi_controller import MPIJobController
+        from libensemble.executors.mpi_controller import MPIJobController
         jobctrl = MPIJobController()
 
     jobctrl.register_calc(full_path=sim_app, calc_type='sim')
