+++ examples/tutorials/tutorial_run_libe_forces.py
@@ -6,7 +6,7 @@ from slim_forces_simf import run_forces  # Sim func from current dir
 from libensemble.libE import libE
 from libensemble.gen_funcs.sampling import uniform_random_sample
 from libensemble.utils import parse_args, add_unique_random_streams
-from libensemble.mpi_controller import MPIJobController
+from libensemble.executors.mpi_controller import MPIJobController
 
 nworkers, is_master, libE_specs, _ = parse_args()  # Convenience function
