+++ libensemble/tests/controller_tests/test_jobcontroller_manager_poll.py
@@ -11,7 +11,7 @@
 # Test does not require running full libensemble
 
 import os
-from libensemble.controller import JobController
+from libensemble.executors.controller import JobController
 
 
 def build_simfunc():
@@ -40,10 +40,10 @@ USE_BALSAM = False  # Take as arg
 
 # Create and add exes to registry
 if USE_BALSAM:
-    from libensemble.balsam_controller import BalsamJobController
+    from libensemble.executors.balsam_controller import BalsamJobController
     jobctrl = BalsamJobController()
 else:
-    from libensemble.mpi_controller import MPIJobController
+    from libensemble.executors.mpi_controller import MPIJobController
     jobctrl = MPIJobController()
 
 jobctrl.register_calc(full_path=sim_app, calc_type='sim')
